requirements.txt
asgiref==3.5.2
bcrypt==4.0.1
cffi==1.15.1
cryptography==38.0.1
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.10.0
django-rest-passwordreset==1.3.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
paramiko==2.11.0
Pillow==9.2.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.3
PyNaCl==1.5.0
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
